コロシアム　たたき台

+++ Assets/Completed/Scripts/Machine_Parameter.cs
@@ -6,7 +6,7 @@ public class Machine_Parameter : MonoBehaviour {
     public float acceleration = 5;//加速度
     public float limmit_speed = 25;//最高速
     public float mass = 1;//重さ
-    public float power = 1;//攻撃力
+    public int power = 1;//攻撃力
     public float friction = 2;//摩擦
